Create common test util mod
